Rename foundation to components
