+++ src/lib/components/LoadingIndicator.svelte
